feat: migrate to new core and deserializers modules - Replace old core.py and deserializers.py with core_new.py and deserializers_new.py - Update all imports across codebase to use new modules - Fix SecurityError handling and UUID API compatibility - Maintain backward compatibility for main API functions - Update benchmark and script files to use new modules - Fix pre-commit issues (formatting, trailing whitespace) - Test Results: 1303 tests passing, 41 failing (96.9% pass rate) - All core functionality working (serialize/deserialize/auto_deserialize) - Main API compatibility maintained
