+++ tests/integration/test_round_trip_serialization.py
@@ -77,7 +77,7 @@ import pytest
 
 import datason
 from datason.config import SerializationConfig
-from datason.deserializers import deserialize_fast
+from datason.deserializers_new import deserialize_fast
 
 # Optional dependencies
 try:
