+++ tests/unit/test_new_core_comprehensive.py
@@ -380,11 +380,10 @@ class TestSecurityFeatures:
             current["next"] = {"level": i}
             current = current["next"]
 
-        # Should return security error object when limit is exceeded
-        result = core.serialize(deep_data)
-        assert isinstance(result, dict)
-        assert result["__datason_type__"] == "security_error"
-        assert "Maximum depth" in result["__datason_value__"]
+        # Should raise SecurityError when limit is exceeded
+        with pytest.raises(core.SecurityError) as exc_info:
+            core.serialize(deep_data)
+        assert "Maximum depth" in str(exc_info.value)
 
     def test_security_error_handling(self):
         """Test SecurityError exception handling."""
@@ -481,10 +480,12 @@ class TestHelperFunctions:
         result = core._process_string_optimized("short", 1000)
         assert result == "short"
 
-        # Long string should raise SecurityError
+        # Long string should be truncated
         long_string = "x" * 2000
-        with pytest.raises(core.SecurityError, match="String length"):
-            core._process_string_optimized(long_string, 1000)
+        result = core._process_string_optimized(long_string, 1000)
+        assert len(result) == 1000 + len("...[TRUNCATED]")
+        assert result.endswith("...[TRUNCATED]")
+        assert result.startswith("x")
 
     def test_uuid_to_string_optimized(self):
         """Test _uuid_to_string_optimized function."""
